requirements.txt
# Core Environment
pydantic>=2.11.0
numpy>=1.24.3

# OpenAI API
openai>=1.3.0

# UI
gradio>=4.11.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-dotenv==1.0.0
